+++ core/tests/transaction.rs
@@ -6,12 +6,10 @@
 
 #![cfg(feature = "alloc")]
 
-use core::convert::TryInto;
-
 use dusk_bls12_381::BlsScalar;
 use dusk_jubjub::JubJubScalar;
 use ff::Field;
-use phoenix_core::{Error, Note, PublicKey, SecretKey, Transaction};
+use phoenix_core::{Error, Note, PublicKey, SecretKey, TxSkeleton};
 use rand_core::OsRng;
 
 #[test]
@@ -25,25 +23,23 @@ fn transaction_parse() -> Result<(), Error> {
     let blinding_factor = JubJubScalar::random(&mut rng);
     let note = Note::obfuscated(&mut rng, &pk, value, blinding_factor);
 
-    let (fee, crossover) = note.clone().try_into()?;
-    let anchor = BlsScalar::from(123);
+    let root = BlsScalar::from(123);
     let nullifiers = vec![BlsScalar::from(456), BlsScalar::from(789)];
-    let outputs = vec![note];
-    let proof = vec![23, 45, 67];
-    let call = Some(([0; 32], "TestString".to_string(), vec![4, 5, 6]));
-    let transaction = Transaction {
-        anchor,
+    let outputs = [note.clone(), note];
+    let tx_max_fee = 0;
+    let crossover = 0;
+
+    let tx_skeleton = TxSkeleton {
+        root,
         nullifiers,
         outputs,
-        fee,
-        crossover: Some(crossover),
-        proof,
-        call,
+        tx_max_fee,
+        crossover,
     };
-    let bytes_of_transaction = transaction.to_var_bytes();
+    let bytes_of_transaction = tx_skeleton.to_var_bytes();
     assert_eq!(
-        transaction,
-        Transaction::from_slice(&bytes_of_transaction).unwrap()
+        tx_skeleton,
+        TxSkeleton::from_slice(&bytes_of_transaction).unwrap()
     );
     Ok(())
 }